+++ HISTORY.md
@@ -6,6 +6,7 @@ v0.8 - 2014/07/18
 * Add content_owner.claims to list the claims administered by a content owner.
 * Allow content_owner.claims to be chained with .where, such as in account.videos.where(q: 'query')
 * Add account.content_owners to list content owners associated with an account
+* Add video.delete
 
 v0.7 - 2014/06/18
 -----------------